Separates graphing from the track control and generation. Separates tests as well
