+++ motbox/__init__.py
@@ -6,5 +6,4 @@ __all__ = ["Track", "Position", "Puppeteer"]
 
 from .track import Track, Position
 from .control import Puppeteer
-from .commands import generate_straight_trajectory
 from .generator import generate_straight_trajectory
