+++ motbox/track.py
@@ -8,12 +8,6 @@ moving objects.
 import numpy as np
 from scipy.spatial import distance
 from scipy.sparse import csgraph
-import matplotlib.pyplot as plt
-
-
-from moviepy.editor import VideoClip
-from moviepy.video.io.bindings import mplfig_to_npimage
-
 
 class Position(object):
     """Represents position of n objects or one timeslice of Track
@@ -109,19 +103,7 @@ class Position(object):
             jitter = np.random.uniform(low=-amount, high=amount, size=(2, self.n_objects))
         self.move((jitter[0, :], jitter[1, :]))
 
-
-    def plot(self, filename, xlim=(-10, 10), ylim=(-10, 10)):
-        """Plots positions into file
-        """
-        plt.figure()
-        plt.plot(self.x, self.y, "o")
-        plt.xlabel = "x"
-        plt.ylabel = "y"
-        plt.xlim(xlim)
-        plt.ylim(ylim)
-        plt.savefig(filename)
-
-
+    
 class Track(object):
     """Track object
 
@@ -258,74 +240,6 @@ class Track(object):
                 self.n_objects, np.amin(self.time), np.amax(self.time))
         else:
             return "Track: NOT initialized"
-
-
-    def plot(self, filename, xlim=(-10, 10), ylim=(-10, 10)):
-        """Plots trajectories into a file. Uses matplotlib
-
-        Parameters
-        -----------
-          filename: str
-            filename for the plot o be saved into
-          xlim: touple of float 
-            x plot limits
-          ylim : touple of float
-
-        Returns
-        -------
-          Saves a png file of name filename into the working directory
-        """
-        plt.figure()
-        plt.plot(self.x, self.y, "k")
-        plt.xlabel = "x"
-        plt.ylabel = "y"
-        plt.xlim(xlim)
-        plt.ylim(ylim)
-        plt.savefig(filename)
-
-
-    def make_video(self, filename, xlim=(-10, 10), ylim=(-10, 10), callback=None, axisOff=True):
-      """Create video of the track and saves it in working directory
-
-      Parameters
-      ---------
-        filename : string 
-          name of the file to be saved into
-        xlim : touple(2) of float
-          xlimits of the fideo
-        ylim : touple(2) of float
-          ylimits of the video
-        callback:
-          what to 
-        axisOff : bool
-
-      Returns
-      --------
-      saved mp4 video at fiven filepath
-      """
-      WIDTH = 900
-      HEIGHT = 600
-      DPI = 150
-      FPS = 25
-      DURATION = np.max(self.time) - np.min(self.time)
-
-      fig, axis = plt.subplots(figsize=(1.0 * WIDTH / DPI, 1.0 * HEIGHT / DPI), dpi=DPI)
-      def make_frame(t):
-          axis.clear()
-          (tx, ty) = self.position_for_time(t + np.min(self.time))
-          axis.plot(tx, ty, "ko")
-          axis.set_xlim(xlim)
-          axis.set_ylim(ylim)
-          axis.set_title("Time {:.2f} s".format(t))
-          if not callback is None:
-              callback(axis)
-          if axisOff:
-              plt.axis('off')
-          return mplfig_to_npimage(fig)
-      animation = VideoClip(make_frame, duration=DURATION)
-      #animation.write_gif(filename, fps=FPS)
-      animation.write_videofile(filename, fps=FPS)
-      pass
 
 
     def bounce_square(self, position, direction, arena_opts):
